Do not execute rm_quotes if ft_glob matched
This is because the resulting strings will exactly be file names and
therefore will not be quoted. If quote characters are in them they are
litteral and are not meant to be interpreted by the shell.

Also, when a file name contains litteral quoting characters (\, ' or ")
and is obtainden via autocompletion, it will have to be quoted to be used
as is by the shell, if not commands wont work as expected.