Handle shell_var rights correctly in export
This adds the possiblity to specify flags to add or remove with
set_shell_var, thus allowing export to work as intended.

Also setting a variable (only_assignment case) now adds the flag
SET to the variable or sets it if the variable does not already
exist. This means that if the value has been previously exported
or created with the export builtin it still is exported (the HOME
value can be changed without export for example).