Fix buffer overflow in parameter_expansions
This was caused by an uncheck return value. When given a bad
substitution such as "${" the shell would assume there was a
closing tag and would therefore read in an unauthorized place
in memory. ft_varlen returns 0 in case the closetag is not
empty and is not found or if the variable name is empty. In all
cases this is an error and the expansion should stop.