Use get_str_qmode instead of is_quote_open
This fixes the echo 'bonjour\' problem where the quote was still open.
However, it also introduces a new problem. if an unquoted line ends in
a '\' it keeps asking for input if no more character is given except
newlines. This is because the last character in the string is still
the '\'.