Fix expansions with quotes inside
The quotes inside expansions (shell variables) must not count like
the quotes typed on input. The solution was to escape them when getting
the value and removing the '\' in rm_quotes. This is a bit hacky, but it
works and an other choice would have meant completely refactoring the
expansions.