Fix tilde expansions
Allow tilde expansions inside assignments and repair the expansions of
"~+" and "~-" which did not work anymore. The only remaining problem is
with inhibitors. The character before the ~ is checked in case of an
assignment. If it is the equal or a ':' then it works, but it should
also work with an assignment like HOME="~".